Restructure project to use `maturin` to build `pyo3` bindings (#88)
