--- .github/workflows/stability.yml
@@ -1,50 +1,0 @@
-name: Code Stability
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-concurrency:
-  group: stability-${{ github.head_ref || github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  stability:
-    name: Code Stability
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Build Python
-        uses: ./.github/actions
-
-      - name: Install tools
-        run: poetry install --only dev
-
-      - id: ruff
-        if: always()
-<<<<<<< HEAD
-        run: poetry run ruff --format=github python_package/surrealdb/
-
-      - id: Black
-        if: always()
-        run: poetry run black python_package/surrealdb/ --check --verbose --diff --color
-
-      - id: mypy
-        if: always()
-        run: poetry run mypy python_package/surrealdb/
-=======
-        run: poetry run ruff check surrealdb/
-
-      - id: Black
-        if: always()
-        run: poetry run black surrealdb/ --check --verbose --diff --color
-
-      - id: mypy
-        if: always()
-        run: poetry run mypy surrealdb/
->>>>>>> 89bdd361d0a92f69deff137958279ab21161f00c
