--- .github/workflows/unit_tests.yml
@@ -1,47 +1,0 @@
-name: unit-tests
-
-on:
-  pull_request:
-  # push:
-    branches:
-      - "*"
-
-jobs:
-  run-unit-tests:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [ "3.7", "3.8", "3.9", "3.10" ]
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Build for tests
-        run: |
-          python tests/scripts/local_build_ci.py
-          pip install docker
-          pip install requests
-#          docker-compose build
-#          docker-compose up -d
-#          sleep 2
-
-      - name: Run Tests
-        run: sh scripts/run_tests.sh
-
-  run-rust-unit-tests:
-    needs: run-unit-tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v2
-  
-      - name: Set up Rust
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-  
-      - name: Build and Run Unit Tests
-        run: cargo test
