+++ Cargo.toml
@@ -12,28 +12,25 @@ once_cell = "1.19.0"
 serde_json = "^1.0.97"
 crossbeam-channel = "^0.5"
 thiserror = "^1.0.43"
+pyo3 = "0.20.0"
+pyo3-asyncio = { version = "0.20",  features = ["unstable-streams", "attributes", "tokio-runtime"] }
+uuid = { version = "1.3.4", features = [ "v4", "fast-rng", "macro-diagnostics" ] }
 
 [dev-dependencies]
 tokio = { version = "1.36.0", features = ["full"] }
 bollard = "*"
 futures-util-preview = "0.2.2"
+pyo3 = { version = "0.20.3", features = ["auto-initialize"] }
 
 [lib]
 name = "rust_surrealdb"
 crate-type=["cdylib"]
 
-[dependencies.pyo3]
-version = "0.20.0"
-features = ["extension-module"]
+[features]
+# must be enabled when building with `cargo build`, maturin enables this automatically
+extension-module = ["pyo3/extension-module"]
 
-[dependencies.pyo3-asyncio]
-version = "0.20"
-features = ["unstable-streams", "attributes", "tokio-runtime"]
-
-[dependencies.uuid]
-version = "1.3.4"
-features = [
-    "v4",                # Lets you generate random UUIDs
-    "fast-rng",          # Use a faster (but still sufficiently random) RNG
-    "macro-diagnostics", # Enable better diagnostics for compile-time UUIDs
-]
+[build-dependencies]
+version_check = "0.9.4"
+# used where logic has to be version/distribution specific, e.g. pypy
+pyo3-build-config = { version = "0.20.2" }
