+++ pyproject.toml
@@ -1,2 +1,37 @@
 [build-system]
-requires = ["setuptools", "wheel", "setuptools-rust"]
+requires = ["maturin>=1.5,<2.0"]
+build-backend = "maturin"
+
+[project]
+name = "surrealdb"
+version = "0.4.0"
+description = "SurrealDB Python Client written in Rust"
+authors = [ 
+    { name = "Maxwell Flitton", email = "maxwellflitton@gmail.com" } 
+]
+readme = "README.md"
+license = { file = "LICENSE" }
+requires-python = ">=3.8"
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Rust",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Programming Language :: Python :: Implementation :: PyPy",
+    "Topic :: Database",
+    "Topic :: Database :: Front-Ends",
+    "Topic :: Software Development",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+keywords = ["SurrealDB", "Database"]
+
+[project.urls]
+repository = "https://github.com/surrealdb/surrealdb.py"
+documentation = "https://surrealdb.com/docs/integration/libraries/python"
+
+[tool.maturin]
+module-name = "surrealdb._rust_surrealdb"
+python-source = "python"
+features = ["pyo3/extension-module"]
