+++ python/surrealdb/__init__.py
@@ -1,0 +1,2 @@
+from .connection_interface import SurrealDB
+from .async_connection_interface import AsyncSurrealDB
