+++ python/surrealdb/async_execution_mixins/create.py
@@ -4,10 +4,10 @@ This file defines the interface between python and the Rust SurrealDB library fo
 import json
 from typing import List, Union
 
-from surrealdb.rust_surrealdb import rust_create_future
-from surrealdb.rust_surrealdb import rust_delete_future
+from .._rust_surrealdb import rust_create_future
+from .._rust_surrealdb import rust_delete_future
 
-from surrealdb.errors import SurrealDbError
+from ..errors import SurrealDbError
 
 
 class AsyncCreateMixin:
