+++ python/surrealdb/async_execution_mixins/query.py
@@ -6,10 +6,10 @@ https://github.com/surrealdb/surrealdb/blob/main/lib/tests/fetch.rs
 import json
 from typing import List, Union
 
-from surrealdb.rust_surrealdb import rust_query_future
-from surrealdb.rust_surrealdb import rust_select_future
+from .._rust_surrealdb import rust_query_future
+from .._rust_surrealdb import rust_select_future
 
-from surrealdb.errors import SurrealDbError
+from ..errors import SurrealDbError
 
 
 class AsyncQueryMixin:
