+++ python/surrealdb/errors.py
