+++ python/surrealdb/execution_mixins/auth.py
@@ -3,12 +3,12 @@ This file defines the interface between python and the Rust SurrealDB library fo
 """
 from typing import Dict, Optional
 
-from surrealdb.rust_surrealdb import rust_authenticate_future
-from surrealdb.rust_surrealdb import rust_sign_in_future
-from surrealdb.rust_surrealdb import rust_sign_up_future
+from .._rust_surrealdb import rust_authenticate_future
+from .._rust_surrealdb import rust_sign_in_future
+from .._rust_surrealdb import rust_sign_up_future
 
-from surrealdb.errors import SurrealDbError
-from surrealdb.asyncio_runtime import AsyncioRuntime
+from ..errors import SurrealDbError
+from ..asyncio_runtime import AsyncioRuntime
 
 
 class SignInMixin:
