+++ python/surrealdb/execution_mixins/update.py
@@ -4,12 +4,12 @@ This file defines the interface between python and the Rust SurrealDB library fo
 import json
 from typing import List, Union
 
-from surrealdb.rust_surrealdb import rust_merge_future
-from surrealdb.rust_surrealdb import rust_update_future
-from surrealdb.rust_surrealdb import rust_patch_future
+from .._rust_surrealdb import rust_merge_future
+from .._rust_surrealdb import rust_update_future
+from .._rust_surrealdb import rust_patch_future
 
-from surrealdb.errors import SurrealDbError
-from surrealdb.asyncio_runtime import AsyncioRuntime 
+from ..errors import SurrealDbError
+from ..asyncio_runtime import AsyncioRuntime
 
 
 class UpdateMixin:
