--- python_package/.editorconfig
@@ -1,20 +1,0 @@
-# EditorConfig coding styles definitions. For more information about the
-# properties used in this file, please see the EditorConfig documentation:
-# http://editorconfig.org/
-
-root = true
-
-[*]
-charset = utf-8
-end_of_line = LF
-indent_size = 4
-indent_style = space
-insert_final_newline = true
-trim_trailing_whitespace = true
-
-[*.{yml,json}]
-indent_size = 2
-indent_style = space
-
-[*.{md,diff}]
-trim_trailing_whitespace = false
