--- python_package/.gitignore
@@ -1,64 +1,0 @@
-# -----------------------------------
-# OS X
-# -----------------------------------
-
-# Directory files
-.DS_Store
-.AppleDouble
-.LSOverride
-
-# Thumbnail files
-._*
-
-# Files that might appear on external disk
-.Spotlight-V100
-.Trashes
-
-# Directories potentially created on remote AFP share
-.AppleDB
-.AppleDesktop
-Network Trash Folder
-Temporary Items
-.apdisk
-
-# Environments
-.env
-.venv
-env/
-venv/
-ENV/
-env.bak/
-venv.bak/
-bin/
-lib64
-pyvenv.cfg
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.nox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.py,cover
-.hypothesis/
-.pytest_cache/
-cover/
-
-# -----------------------------------
-# Files
-# -----------------------------------
-
-*.db
-*.pyc
-.vscode/
-__pycache__/
-dist/
-
-# -----------------------------------
-# IDE
-# -----------------------------------
-.idea/
