--- python_package/.pre-commit-config.yaml
@@ -1,20 +1,0 @@
-default_language_version:
-  python: python3
-
-repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
-    hooks:
-    -   id: check-ast
-    -   id: check-yaml
-    -   id: end-of-file-fixer
-    -   id: requirements-txt-fixer
-    -   id: trailing-whitespace
-- repo: https://github.com/charliermarsh/ruff-pre-commit
-  rev: 'v0.0.245'
-  hooks:
-    - id: ruff
--   repo: https://github.com/psf/black
-    rev: 22.8.0
-    hooks:
-    -   id: black
