--- python_package/README.md
@@ -1,135 +1,0 @@
-<p align="center">
-    <img width="300" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/icon.png" alt="SurrealDB Icon">
-</p>
-
-<br>
-
-<p align="center">
-    <a href="https://surrealdb.com#gh-dark-mode-only" target="_blank">
-        <img width="300" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/white/logo.svg" alt="SurrealDB Logo">
-    </a>
-    <a href="https://surrealdb.com#gh-light-mode-only" target="_blank">
-        <img width="300" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/black/logo.svg" alt="SurrealDB Logo">
-    </a>
-</p>
-
-<h3 align="center">
-    <a href="https://surrealdb.com#gh-dark-mode-only" target="_blank">
-        <img src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/white/text.svg" height="15" alt="SurrealDB">
-    </a>
-    <a href="https://surrealdb.com#gh-light-mode-only" target="_blank">
-        <img src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/black/text.svg" height="15" alt="SurrealDB">
-    </a>
-    is the ultimate cloud <br> database for tomorrow's applications
-</h3>
-
-<h3 align="center">Develop easier. &nbsp; Build faster. &nbsp; Scale quicker.</h3>
-
-<br>
-
-<p align="center">
-    <a href="https://github.com/surrealdb/surrealdb.py"><img src="https://img.shields.io/badge/status-beta-ff00bb.svg?style=flat-square"></a>
-    &nbsp;
-    <a href="https://surrealdb.com/docs/integration/libraries/python"><img src="https://img.shields.io/badge/docs-view-44cc11.svg?style=flat-square"></a>
-    &nbsp;
-    <a href="https://github.com/surrealdb/surrealdb.py"><img src="https://img.shields.io/badge/license-Apache_License_2.0-00bfff.svg?style=flat-square"></a>
-    &nbsp;
-    <a href="https://twitter.com/surrealdb"><img src="https://img.shields.io/badge/twitter-follow_us-1d9bf0.svg?style=flat-square"></a>
-    &nbsp;
-    <a href="https://dev.to/surrealdb"><img src="https://img.shields.io/badge/dev-join_us-86f7b7.svg?style=flat-square"></a>
-    &nbsp;
-    <a href="https://www.linkedin.com/company/surrealdb/"><img src="https://img.shields.io/badge/linkedin-connect_with_us-0a66c2.svg?style=flat-square"></a>
-</p>
-
-<p align="center">
-	<a href="https://surrealdb.com/blog"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/blog.svg" alt="Blog"></a>
-	&nbsp;
-	<a href="https://github.com/surrealdb/surrealdb"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/github.svg" alt="Github	"></a>
-	&nbsp;
-    <a href="https://www.linkedin.com/company/surrealdb/"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/linkedin.svg" alt="LinkedIn"></a>
-    &nbsp;
-    <a href="https://twitter.com/surrealdb"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/twitter.svg" alt="Twitter"></a>
-    &nbsp;
-    <a href="https://www.youtube.com/channel/UCjf2teVEuYVvvVC-gFZNq6w"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/youtube.svg" alt="Youtube"></a>
-    &nbsp;
-    <a href="https://dev.to/surrealdb"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/dev.svg" alt="Dev"></a>
-    &nbsp;
-    <a href="https://surrealdb.com/discord"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/discord.svg" alt="Discord"></a>
-    &nbsp;
-    <a href="https://stackoverflow.com/questions/tagged/surrealdb"><img height="25" src="https://raw.githubusercontent.com/surrealdb/surrealdb/main/img/social/stack-overflow.svg" alt="StackOverflow"></a>
-
-</p>
-
-# surrealdb.py
-
-The official SurrealDB library for Python.
-
-[See the documentation here](https://surrealdb.com/docs/integration/libraries/python) 
-
-## Getting Started
-
-First install [SurrealDB](https://surrealdb.com/docs/start/installation) if you haven't already.
-
-### Install the python library
-```
-pip install surrealdb
-```
-
-Alternatively, you can use install it using [Poetry](https://python-poetry.org/)
-
-```python
-from surrealdb import Surreal
-
-async def main():
-    """Example of how to use the SurrealDB client."""
-    async with Surreal("ws://localhost:8000/rpc") as db:
-        await db.signin({"user": "root", "pass": "root"})
-        await db.use("test", "test")
-        await db.create(
-            "person",
-            {
-                "user": "me",
-                "pass": "safe",
-                "marketing": True,
-                "tags": ["python", "documentation"],
-            },
-        )
-        print(await db.select("person"))
-        print(await db.update("person", {
-            "user":"you",
-            "pass":"very_safe",
-            "marketing": False,
-            "tags": ["Awesome"]
-        }))
-        print(await db.delete("person"))
-
-        # You can also use the query method 
-        # doing all of the above and more in SurrealQl
-        
-        # In SurrealQL you can do a direct insert 
-        # and the table will be created if it doesn't exist
-        await db.query("""
-        insert into person {
-            user: 'me',
-            pass: 'very_safe',
-            tags: ['python', 'documentation']
-        };
-        
-        """)
-        print(await db.query("select * from person"))
-        
-        print(await db.query("""
-        update person content {
-            user: 'you',
-            pass: 'more_safe',
-            tags: ['awesome']
-        };
-        
-        """))
-        print(await db.query("delete person"))
-
-if __name__ == "__main__":
-    import asyncio
-
-    asyncio.run(main())
-```
