--- python_package/SECURITY.md
@@ -1,12 +1,0 @@
-# Security Policy
-
-## Reporting a Vulnerability
-
-We take the security of SurrealDB code, software, and cloud platform very
-seriously. If you believe you have found a security vulnerability in
-SurrealDB, we encourage you to let us know right away. We will investigate
-all legitimate reports and do our best to quickly fix the problem.
-
-Please report any issues or vulnerabilities to security@surrealdb.com,
-instead of posting a public issue in GitHub. Please include the version
-identifier, and details on how the vulnerability can be exploited.
