+++ src/lib.rs
@@ -34,6 +34,7 @@ macro_rules! py_future_wrapper {
 
 
 #[pymodule]
+#[pyo3(name="_rust_surrealdb")]
 fn rust_surrealdb(_py: Python, m: &PyModule) -> PyResult<()> {
     let _ = m.add_wrapped(wrap_pyfunction!(rust_make_connection_future));
     let _ = m.add_wrapped(wrap_pyfunction!(rust_sign_in_future));
