--- surrealdb/VERSION.txt
@@ -1,1 +1,0 @@
-VERSION='0.0.2'
