--- surrealdb/__init__.py
@@ -1,2 +1,0 @@
-from surrealdb.connection_interface import SurrealDB
-from surrealdb.async_connection_interface import AsyncSurrealDB
